chage model_inference.py to api like version
