+++ pdllib/task_map.py
@@ -1,0 +1,32 @@
+task_map = {
+    "promoters": {
+        "labels": ["Not promoter", "Core promoter"]
+    },
+    "H3K27ac": {
+        "labels": ["Not H3K27ac", "H3K27ac"]
+    },
+    "H3K27me3": {
+        "labels": ["Not H3K27me3", "H3K27me3"]
+    },
+    "H3K4me3": {
+        "labels": ["Not H3K4me3", "H3K4me3"]
+    },
+    "DNA_methylation": {
+        "labels": ["Not methylated", "Methylated"]
+    },
+    "conservation": {
+        "labels": ["Not conserved", "Conserved"]
+    },
+    "lncRNAs": {
+        "labels": ["Not lncRNA", "lncRNA"]
+    },
+    "open_chromatin": {
+        "labels": ["Not open chromatin", "Full open chromatin", "Partial open chromatin"]
+    },
+    "pro_str_leaf": {
+        "labels": ["Promoter strength in tobacco leaves"]
+    },
+    "pro_str_protoplast": {
+        "labels": ["Promoter strength in maize protoplasts"]
+    }
+}
